+++ src/day02.rs
@@ -5,12 +5,12 @@ use crate::solution::Solution;
 pub struct Day;
 
 impl Solution for Day {
-    fn part1(&self, input: &str) -> Option<u64> {
+    fn part1(&self, input: &str) -> Option<usize> {
         let games = input.lines().map(|s| s.parse::<Game>().unwrap());
         Some(games.filter(|game| game.is_possible((12, 13, 14))).map(|game| game.id).sum())
     }
 
-    fn part2(&self, input: &str) -> Option<u64> {
+    fn part2(&self, input: &str) -> Option<usize> {
         let games = input.lines().map(|s| s.parse::<Game>().unwrap());
         Some(games.map(|game| game.get_power()).sum())
     }
@@ -18,12 +18,12 @@ impl Solution for Day {
 
 #[derive(Debug)]
 struct Game {
-    id: u64,
-    draws: Vec<(u64, u64, u64)>,
+    id: usize,
+    draws: Vec<(usize, usize, usize)>,
 }
 
 impl Game {
-    pub fn is_possible(&self, max: (u64, u64, u64)) -> bool {
+    pub fn is_possible(&self, max: (usize, usize, usize)) -> bool {
         for draw in &self.draws {
             if draw.0 > max.0 || draw.1 > max.1 || draw.2 > max.2 {
                 return false;
@@ -32,7 +32,7 @@ impl Game {
         true
     }
 
-    pub fn get_power(&self) -> u64 {
+    pub fn get_power(&self) -> usize {
         let mut max = (0, 0, 0);
         for draw in &self.draws {
             if draw.0 > max.0 {
@@ -54,7 +54,7 @@ impl FromStr for Game {
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let (id, bag_draws) = s.split_once(':').unwrap();
-        let id = id.split_once(' ').unwrap().1.parse::<u64>().unwrap();
+        let id = id.split_once(' ').unwrap().1.parse::<usize>().unwrap();
         let draws = bag_draws
             .split(';')
             .map(|s| s.trim())
@@ -64,9 +64,9 @@ impl FromStr for Game {
                 for pull in pulls {
                     let (count, color) = pull.split_once(' ').unwrap();
                     match color {
-                        "red" => draw.0 = count.parse::<u64>().unwrap(),
-                        "green" => draw.1 = count.parse::<u64>().unwrap(),
-                        "blue" => draw.2 = count.parse::<u64>().unwrap(),
+                        "red" => draw.0 = count.parse::<usize>().unwrap(),
+                        "green" => draw.1 = count.parse::<usize>().unwrap(),
+                        "blue" => draw.2 = count.parse::<usize>().unwrap(),
                         _ => panic!("Unknown color {}", color),
                     }
                 }
